pyproject.toml
[tool.poetry]
name = "toggl-round"
version = "0.1.0"
description = ""
authors = ["Søren Siersted Mikkelsen <soren@aamikkelsen.dk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sniffer = "^0.4.1"
rollbar = "^1.0.0"
requests = "^2.31.0"
pytz = "^2024.1"
python-termstyle = "^0.1.10"
python-dateutil = "^2.9.0.post0"
pytest = "^8.0.2"
py = "^1.11.0"
nose = "^1.3.7"
colorama = "^0.4.6"
black = "^24.2.0"
isort = "^5.13.2"
pylsp-mypy = "^0.6.8"
pylsp-rope = "^0.1.11"
python-lsp-ruff = "^2.2.0"
types-pytz = "^2024.1.0.20240203"
types-requests = "^2.31.0.20240218"
types-python-dateutil = "^2.8.19.20240106"
loguru = "^0.7.2"
ipython = "^8.22.2"
ipdb = "^0.13.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
